Solver specification made more flexible

+++ src/gnome/gronor_mods.F90
@@ -884,7 +884,7 @@ end module mkl_solver
 
 #ifdef LAPACK
 module lapack_solver
-  integer (kind=8) :: lwork1m,lwork2m,lwork,lworki,liwork,ndimm,mdimm
+  integer (kind=4) :: lwork1m,lwork2m,lwork,lworki,liwork,ndimm,mdimm
   real(kind=8), allocatable :: work(:)
   integer(kind=8), allocatable :: iwork(:)
   real (kind=8),allocatable :: workspace_d(:)
